refactor: move all public types to types/index.d.ts
