+++ packages/vuetify/src/components/VCalendar/VCalendar.ts
@@ -30,7 +30,7 @@ import {
 import VCalendarMonthly from './VCalendarMonthly'
 import VCalendarDaily from './VCalendarDaily'
 import VCalendarWeekly from './VCalendarWeekly'
-import { VTimestamp, VTime } from './types'
+import { VTimestamp, VTime } from 'types'
 
 // Types
 interface VCalendarRenderProps {
