+++ packages/vuetify/src/components/VCalendar/mixins/calendar-base.ts
@@ -20,7 +20,7 @@ import {
   getStartOfWeek,
   getEndOfWeek,
 } from '../util/timestamp'
-import { VTimestamp, VTimestampFormatter } from '../types'
+import { VTimestamp, VTimestampFormatter } from 'types'
 
 export default mixins(
   Colorable,
