+++ packages/vuetify/src/components/VCalendar/util/events.ts
@@ -3,7 +3,7 @@ import {
   getDayIdentifier,
   getTimestampIdentifier,
 } from './timestamp'
-import { VTimestamp, VEventInput } from '../types'
+import { VTimestamp, VEventInput } from 'types'
 
 export interface VEventParsed {
   input: VEventInput
