+++ packages/vuetify/src/components/VDatePicker/VDatePickerDateTable.ts
@@ -8,7 +8,7 @@ import mixins from '../../util/mixins'
 
 // Types
 import { VNode, VNodeChildren, PropType } from 'vue'
-import { DatePickerFormatter } from './types'
+import { DatePickerFormatter } from 'types'
 
 export default mixins(
   DatePickerTable
