+++ packages/vuetify/src/components/VDatePicker/VDatePickerYears.ts
@@ -10,7 +10,7 @@ import mixins, { ExtractVue } from '../../util/mixins'
 
 // Types
 import Vue, { VNode, PropType } from 'vue'
-import { DatePickerFormatter } from './types'
+import { DatePickerFormatter } from 'types'
 
 interface options extends Vue {
   $el: HTMLElement
