--- packages/vuetify/src/components/VSelect/types.d.ts
@@ -1,1 +1,0 @@
-export type ItemKey = string | (string | number)[] | ((item: object, fallback?: any) => any)
