--- packages/vuetify/src/components/VTreeview/types.d.ts
@@ -1,1 +1,0 @@
-export type FilterTreeItemFunction = (item: object, search: string, textKey: string) => boolean
