--- packages/vuetify/src/directives/touch/types.d.ts
@@ -1,25 +1,0 @@
-export interface TouchHandlers {
-  start?: (wrapperEvent: TouchEvent & TouchWrapper) => void
-  end?: (wrapperEvent: TouchEvent & TouchWrapper) => void
-  move?: (wrapperEvent: TouchEvent & TouchWrapper) => void
-  left?: (wrapper: TouchWrapper) => void
-  right?: (wrapper: TouchWrapper) => void
-  up?: (wrapper: TouchWrapper) => void
-  down?: (wrapper: TouchWrapper) => void
-}
-
-export interface TouchWrapper extends TouchHandlers {
-  touchstartX: number
-  touchstartY: number
-  touchmoveX: number
-  touchmoveY: number
-  touchendX: number
-  touchendY: number
-  offsetX: number
-  offsetY: number
-}
-
-export type TouchValues = TouchHandlers & {
-  parent?: boolean
-  options?: AddEventListenerOptions
-}
