--- packages/vuetify/src/mixins/validatable/types.d.ts
@@ -1,5 +1,0 @@
-export type VuetifyRuleValidator = (value: any) => string | boolean
-
-export type VuetifyMessage = string | string[]
-
-export type VuetifyRuleValidations = (VuetifyRuleValidator | string)[]
